GCP/1.4.1 Telegram Bot/requirements.txt
flask
google-cloud-firestore
google-cloud-pubsub
gunicorn
numpy